# internal/api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: pool.biswas.me
  version: "1.0"
  summary: Pool water chemistry, treatment planning and cost-of-ownership tracking.
  description: |
    Every action the web interface performs is one of these calls — there is no
    separate API surface to keep in sync.

    **Authentication.** Send a personal API key as `Authorization: Bearer
    pool_sk_…`, or as `X-API-Key`. Create one under Settings → API keys in the
    app; the plaintext is shown once and cannot be recovered. A key carries a
    `read` or `read,write` scope, enforced on the HTTP method: a read key
    cannot POST, PATCH or DELETE. Browser sessions authenticate the same
    endpoints with a cookie.

    **Chemistry.** Readings are stored as given. Combined chlorine, adjusted
    alkalinity, the Langelier saturation index and the 0–100 water quality
    score are computed server-side from the pool's volume, surface and
    sanitizer, so a test posted through the API is scored identically to one
    typed into the app.

    **For an agent doing the reading.** The photo path is a convenience, not the
    only way in. An agent with a stronger model can do the whole job through
    the ordinary endpoints: `POST /api/tests` with the readings it transcribed,
    `POST /api/attachments` with the photo and that `test_id`, then
    `GET /api/tests/{id}/context` for the brief and `POST
    /api/tests/{id}/insight` with the analysis it wrote. Nothing about that
...
    say what did the reading.

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKeyHeader: []
...
paths:
...
  /api/me/passkeys/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    patch:
      tags: [Account]
      summary: Rename a passkey
      responses:
        "200": { description: Renamed. }
    delete:
      tags: [Account]
      summary: Remove a passkey
      responses:
        "200": { description: Removed. }
...
  /api/keys:
    get:
      tags: [Account]
      summary: List your API keys
      description: Secrets are never returned; a key is identified by its prefix.
      responses:
        "200":
          description: Your keys, newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/APIKey" }
    post:
      tags: [Account]
      summary: Issue an API key
      description: |
        The plaintext key is in this response and nowhere else, ever again.
        The response also carries the OpenAPI document's location and a `curl`
        that works as printed.

        A key cannot issue another key: that would let a read credential
        quietly escalate itself into a write one. Sign in to create one.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80, examples: ["Home Assistant"] }
                scopes:
                  description: "`read`, or `read,write`. Anything unrecognised normalises to read."
                  oneOf:
                    - type: string
                      examples: ["read,write"]
                    - type: array
                      items: { type: string, enum: [read, write] }
                expires_at:
                  type: string
                  description: RFC3339. Omit for a key that never expires.
      responses:
        "201":
          description: The key, shown once.
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    allOf:
                      - $ref: "#/components/schemas/APIKey"
                      - type: object
                        properties:
                          key:
                            type: string
                            description: The plaintext. Store it now.
                            examples: ["pool_sk_9f3c1a20…"]
                  discovery:
                    type: object
                    properties:
                      base_url: { type: string }
                      spec_url: { type: string }
                      auth_scheme: { type: string, examples: [Bearer] }
                      example: { type: string }
                      scopes: { type: array, items: { type: string } }
        "403": { description: An API key cannot issue another key. }
  /api/keys/{id}:
    delete:
      tags: [Account]
      summary: Revoke an API key
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Revoked. }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer pool_sk_…`"
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
  responses:
...
    NotFound:
      description: No such record on this account.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    APIKey:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        prefix:
          type: string
          description: The identifying, non-secret leading portion.
          examples: ["pool_sk_9f3c1a20"]
        scopes: { type: string, examples: ["read,write"] }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string }
        expires_at: { type: string }
        revoked_at: { type: string }
...